refactor: UpperTag解耦合
